requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
decorator==5.1.1
distro==1.9.0
executing==2.0.1
Flask==3.0.3
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
parso==0.8.4
pdfminer.six==20240706
pexpect==4.9.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
Werkzeug==3.0.3